refactor: ♻️ Change "insecure" config setting to verify and tweak logic
I changed the "insecure" config setting to instead be whether we should verify the certificate or not. This allows us to use it as the variable for the authentication call and logically makes a bit more sense than the way it was done before. The setting before was acting in opposite behavior which is quite different than how this issue is typically handled.