refactor: ♻️ Refactor to have a CloudvisionApi for connections to CVP
Updated the CVP functions to be part of CloudvisionApi class and updated Job to have it passed in. This will allow for easier testing in the future and to address a bunch of pylint complaints for the global and other choices that using OOP resolves.

Moved functions for communicating to CVP inside an Api class.