Project Refactor to New Pattern (#85)
* refactor: ♻️ Move util functions to appropriate folder

To follow django organizational standards utility functions should be in a utils folder. I've also renamed the files to be the system of record to make the name space more sensible.

Moved utility functions to a new path.

* refactor: ♻️ Refactor project to update structure to standardized pattern

I've updated the project structure to follow our current guidelines for SSoT plugins. This should allow for easier maintenance and functional updates.

Multiple classes have been renamed, files moved around and various other changes that will break outside packages.

* fix: Correct CustomField loading

* test: 🚨 Ensure Exception is called properly

* refactor: ♻️ Refactor to have a CloudvisionApi for connections to CVP

Updated the CVP functions to be part of CloudvisionApi class and updated Job to have it passed in. This will allow for easier testing in the future and to address a bunch of pylint complaints for the global and other choices that using OOP resolves.

Moved functions for communicating to CVP inside an Api class.

* test: 🚨 Fix various linter complaints

* chore: 🔥 Remove SSoT 0.1.0 framework package as not needed any longer

The framework is now publically available and is at 1.0.0+ at this point so this file shouldn't be needed anymore.

* docs: 🔖 Bump to version 1.2.1

* fix: Make Nautobot 1.2.0 minimum supported version

Nautobot 1.2.0 is minimum supported version now.

* docs: 📝 Update CHANGELOG

* fix: Require device model be defined for Device

* fix: 🐛 Remove Diffsync var from update, it's not passed

* fix: 🐛 Correct create func to use attr

Not sure why a call to CVP was made when the information is passed. This removes the unnecessary API call and uses the information passed to the function.

* fix: ♻️ Correct CVP CRUD ops to use CloudvisionApi

This should get the functionality operational again but this entire portion of the project needs to be refactored to remove the calls to CVP. The information being passed is information that should be stored in the DiffSync model and passed to the CRUD funcs as needed.

* test: 🚨 Remove unused import for linter

* docs: 📝 Update CHANGELOG to add part for 1.2.0

Co-authored-by: Uros Bajzelj <ubajze@users.noreply.github.com>
Co-authored-by: Justin Drew <jdrew82@users.noreply.github.com>