refactor: ♻️ Update CloudvisionApi to match GRPCClient from Arista library
This is done to test the example methods and hopefully utilize for getting interface data.

Got rid of disconnect method and using close now. Also updated to work as context manager and to automatically do connection in dunder init instead of connect method. Also have moved get functions to be outside class.