Import Interfaces (#94)
* feat: ✨ Add serial attribute to Device

Store DeviceID as serial for Device. Updated adapters to load that data into model and updated the CRUD funcs to import that data into Nautobot.

* fix: 🔥 Remove CF assignment functions as not needed

Found these functions in use for assigning default CustomFields to devices. This isn't necessary as once CustomFields are created and assigned to Devices they'll show up for all that are created.

* refactor: ♻️ Make connection for CloudvisionAdapter required, specify type

* feat: ✨ Add Port DiffSyncModel for interface syncing

* build: ⬆️ Set cloudvision to 1.4.5 for GRPCClient changes

We need to use Cloudvision library v1.4.5 in order to be able to pass the string value of the API endpoint's certificate instead of having to pass a file for the GRPCClient. For some reason 1.4.5 isn't in PyPI so I'm pulling from their git instead.

* refactor: ♻️ Store API endpoint cert in var for GRPCClient

Changed operation so that it always stores the endpoint certificate as it's required for the GRPCClient class when connecting to a device directly for information.

* feat: ♻️ Tweak attrs on Port, add Port as child to Device

* fix: 🐛 Change cvp_cert to note be bytes encoded

* build: ⬆️ Upgrade cloudvision dep to 1.5.0

* fix: 🐛 Add exception handling for duplicate device hostnames

* fix: Set cvp_token to sessionID if not defined already

* feat: Add GRPCClient from example

* fix: Set token to sessionID if it is blank

* build: ⬆️ Upgrade to cloudvision 1.5.0

* refactor: ♻️ Update CloudvisionApi to match GRPCClient from Arista library

This is done to test the example methods and hopefully utilize for getting interface data.

Got rid of disconnect method and using close now. Also updated to work as context manager and to automatically do connection in dunder init instead of connect method. Also have moved get functions to be outside class.

* refactor: 🚚 Move get functions outside class

Get functions are now called directly instead of through class and now require client to be passed.

* build: Update dependencies

* fix: Update methods to not come from Api class

* refactor: ♻️ Use CloudvisionApi as context manager in Jobs

* fix: 🐛 Ensure client is the comm_channel from CVP for device methods

* fix: 🐛 Correct call to CVP to include client

* style: 🚨 Fix formatting for black

* feat: ✨ Update methods to get all interface information

* feat: ✨ Add method to get interface transceiver type from gRPC

In order to create ports we need to know what type they are to create Interfaces in Nautobot.

* feat: ✨ Add enabled attribute to Port

* docs: 📝 Fix docstrings and add type hints

* feat: ✨ Update adapter load method for ports, add CRUD methods for ports

* refactor: 🔥 Remove unused import

* test: 🚨 Fix docstrings for pydocstring

* test: 🚨 Fixes for pylint

* feat: ✨ Add PORT_TYPE_MAP constant for mapping xcvr to Nautobot type

* feat: Add mode attr to Interface base model

* refactor: 🚚 Rename type to port_type to avoid conflict with reserved word, remove speed

Removed speed attr as it's not needed to figure out the port type.

* refactor: 🔥 Removed speed and duplex from interface data, added operational status

Changed the keys for the interface dict being returned to specify the link or operational status.

* docs: 📝 Fix docstring to be more accurate

* fix: 🐛 Changed the keys looked for when determining transceiver

This is being done as actualldEEepromContents appears to always be there where the detect wasn't on all models. This also adds handling for embedded interfaces on certain models.

* feat: Add method to get interface mode

* feat: Add method to determine port type

* feat: Add method get status from link/oper status

* feat: Interface add mode update,add status to both

* fix: ♻️ Update adapters to address changes in Port model

* fix: Correct type hint for method

* test: 🚨 Fix for pylint

* fix: 🐛 Correct client being passed to self.conn itself

* fix: Correct device name attribute for port load

* refactor: ♻️ Make CFs a top level object

In order to ensure ports are processed we need to make CFs their own object and not related to a Device.

* docs: 🔊 Add debug logging for port import

* docs: 🔊 Add some debug logging for port import

* fix: Correct key to be enabledState

* refactor: 🔥 Remove diffsync from get_query as it's not useful and doesn't print to log

* docs: 🔊 Update log to be clearer about error.

* fix: Remove diffsync from method calls

* fix: Correct key to intfId for interface name

* fix: Add check for blank interface name

* fix: Correct port status to call for object

* fix: Correct key for transceiver type

* fix: Specify port type when creating, fix update

* fix: Catch validation errors when creating devices

* fix: Correct slug for decomm intf status

* fix: Tweak get intf to handle missing keys

* fix: Add checks for possible missing keys

* fix: Check key exists, update logging

* fix: 🐛 Check for intfId key to filter port import

Found that gRPC returns a lot of unnamed ports that need to be filtered out.

* test: 🚨 Remove unused import for flake8

* test: ✅ Add test for get_devices method

* build: ➕ Add parameterized as dev dependency

* test: ✅ Add test for get_port_type method

* test: ✅ Add test for get_interface_status method

* style: 🚨 Fix import sort order

* fix: ♻️ Check base interface for transceiver

In cases where a breakout SFP is used, ie a 40G -> 4x10G, we need to look at the base interface to get the transceiver information. This should hopefully have more interfaces showing the actual transceiver instead of Unknown.

* test: ✅ Add test for get_tags method.

* fix: Correct base port to use interface key

* fix: Check for mediaType key

* docs: 📝 Fix typo in docstrings

* refactor: 🔥 Remove ValidationError catch, it shouldn't be thrown during load

* test: ✅ Add test for get_interfaces_fixed method

* refactor: 🚚 Move fixtures to own file

* test: ✅ Add initial tests for device and interface import in CVP adapter

* refactor: ♻️ Refactored load method into submethods for each object for tests

To make the unit tests easier I broke up the original load method into individual methods for each object, device, interface, and tags. As they're related the load method now only calls the load_device which then gets interfaces and tags.

* docs: 📝 Fix README grammar, add notes about interfaces, adjust for var and functionality changes

* fix: Fix uuid to be id for loading NB Interfaces

* refactor: ♻️ Catch ValidationError for device update

Updating the device update function to catch a ValidationError on updating a Device, just in case.

* docs: 📝 Tweak README intro

* fix: 🐛 Catch possible blank mac addr; should be blank string

Co-authored-by: Justin Drew <jdrew82@users.noreply.github.com>