fix: ♻️ Redo how Nautobot adapter load is performed. Fixes issues #93 and #95.
I found that in an update run that the Nautobot adapter was returning no Devices. In my investigation it appears it was being caught in the try/except block for some AttributeError. I've refactored this to narrow down the block to just CustomField loading and have some logging if encountered.