Add support for Device Lifecycle Plugin (#101)
* fix: 🐛 Fix Platform to be created in signals, remove from CustomField create/update

This should fix Issue #100. The Platform should not be the device model. It should be one for all Device Types and should be the Arista EOS platform. This fixes documentation and removes the errneous creation/update from tags.

* build: ➕ Add Device Lifecycle Plug-in as optional dependency to enable support

* feat: ✨ Add version to Device, enable support for DLC plugin

* test: 🚨 Note unused import for linters

* refactor: ♻️ Set manufacturer for new Software versions to Arista

* fix: Correct filter to device_platform

* fix: 🐛 Remove Manufacturer from SoftwareLCM

Meant to put this on Platform.

* refactor: ♻️ Update Manufacturer to be created in signal, add to Platform creation

Also removed verify_manufacturer method as it's no longer needed.

* build: 🔨 Add DLC plugin to dev environment build

* refactor: 🚚 Move Manufacturer and Platform creation to own signal methods

* build: 🐛 Fix step where DLC plugin is installed

* build: Update project dependencies

* build: Add DLC plugin to PLUGINS

* fix: Correct DeviceType route link in mappings

* fix: Ensure MAC address is string when loaded

* fix: ♻️ Redo how Nautobot adapter load is performed. Fixes issues #93 and #95.

I found that in an update run that the Nautobot adapter was returning no Devices. In my investigation it appears it was being caught in the try/except block for some AttributeError. I've refactored this to narrow down the block to just CustomField loading and have some logging if encountered.

* fix: Correct device_type name attribute to model

* fix: 🐛 Correct method to startswith, correct CF for version

* fix: 🐛 Correct indentation on arista_model CF load

The indentation was wrong causing the CF to be loaded repeatedly causing an ObjectAlreadyExists to be thrown.

* fix: 🐛 Remove arista_model in load/delete op; should have done with Platform fix

* fix: 🐛 Check for enabledState defined, default False for enabled, fixes #102

* fix: 🐛 Ensure MAC addr is lowercase to match CVP for diff

* docs: 📝 Add not about DLC support for EOS Version

Co-authored-by: Justin Drew <jdrew82@users.noreply.github.com>