fix: 🐛 Add handling for unknown device type. Fixes #116
It was discovered that APs or devices onboarded through SNMP return a blank dict for device type query. This will handle the missing data and raise an error in the logs noting that the device interfaces won't be processed. Also added a unit test for the knew device type response.