Fix Hostname Parsing (#122)
* test: ✅ Update override settings for tests to be explicit to remove dependence on dev config.

To ensure that any changes to the plugin settings in the dev settings doesn't impact the test I'm explicitly setting the hostname patterns and site or role mappings accordingly.

* fix: 🐛 Add validation that site or role named capture group is in hostname_patterns. Fixes 119.

To ensure that the match group for site or role is defined in the hostname_patterns setting we need to check for it. This fixes the issue that was detailed in #119. I've also included some unit tests to confirm that defining only site or role capture groups should work.

Co-authored-by: Justin Drew <jdrew82@users.noreply.github.com>